% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schaefer.R
\name{calc.MSY}
\alias{calc.MSY}
\title{Calculating Schaefer's management parameters}
\usage{
calc.MSY(
  K,
  B0,
  r,
  q,
  s.sigma,
  df,
  OWT = FALSE,
  currentF = 0.7,
  weight = 0.9,
  plot = FALSE
)
}
\arguments{
\item{K}{surplus production parameter which represents carrying capacity}

\item{B0}{surplus production parameter which represents biomass when no fishing
activity has started}

\item{r}{surplus production parameter which represents intrinsic growth rate}

\item{q}{surplus production parameter which represents catchability coefficient}

\item{s.sigma}{surplus production parameter which represents observation error}

\item{df}{dataframe containing three columns; year, catch and effort. A fourth column with biomass should be added
if OWT (One Way Trip) option uses "Biomass"}

\item{OWT}{is CPUE plot showing One Way Trip pattern? The default is FALSE, but should be
replaced with either "Biomass" or "Depletion" when the plot shows One Way Trip type of data}

\item{currentF}{Current exploitation rate collected from other survey. Only being used
when OWT="Depletion" is chosen}

\item{weight}{weight given to the deviation between observed and predicted value in either
biomass or exploitation rate}

\item{plot}{option to show the plot as result of fitted estimation to observation data}
}
\value{
A penalized likelihood is used to fix the lack of contrast in One Way Trip type of data using Depletion or Biomass data.

The Biomass option in OWT is used when biomass time series data from acoustic or trawl survey is available and
should be added as the fourth columns in the input dataframe. The default weight when Biomass level is set at 0.9 with range
between 0-1 (lower accuracy with high variance as closer to 0, constrain the estimation procedure to fit the auxiliary
information as closer to 1)

The Depletion option in OWT uses current harvest rate from survey or expert knowledge as penalty.
Depletion range is between 0 to 1, where higher number represent higher depletion level. The default is 0.7 to
say that the depletion is high and many fish were caught. The default weight for harvest rate is 1000 and can be adjusted
so the predicted harvest rate reach a closest value to the current exploitation rate. Predicted harvest rate value
in each optimization step will show up when optimization process is being executed.
}
\description{
This function calculates Schaefer's management parameters using maximum likelihood minimization.
Observation error is used to increase the accuracy of data fitting. It is assumed to occur in the relationship
between stock biomass and index of abundance and is estimated assuming lognormal distribution in
maximum likelihood (Polacheck et al., 1993).

Since fishing effort data collection are not always conducted regularly while catch is likely
have a better time series information, this function also allow for some lose of catch and effort data.

This function also consider the different quality data, for instance if the data shows
a one way trip pattern which losing rate of catch increase.
}
\examples{

calc.MSY(K=1000, B0=1000, r=0.2, q=0.00025, s.sigma=0.1, df=df.goodcontrast,
OWT=FALSE, currentF = 0.7, weight=0.5, plot=TRUE)

}
\references{
Hilborn, Ray, and Carl J. Walters, eds. Quantitative fisheries stock assessment: choice,
dynamics and uncertainty. Springer Science & Business Media, 1992.

Polacheck, T., Hilborn, R., and A.E. Punt. 1993. Fitting surplus production models:
Comparing methods and measuring uncertainty. Canadian Journal of Fisheries and Aquatic Sciences, 50: 2597-2607.
}
