% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_eastpacCatch.R
\docType{data}
\name{df.eastpacCatch}
\alias{df.eastpacCatch}
\title{pelagic longline fishery data}
\format{
A data frame with 21 rows and 3 variables:
\describe{
  \item{year}{year when fishing activity occured.}
  \item{catch}{catch per annum in MT.}
  \item{effort}{effort unit per annum.}
}
}
\usage{
df.eastpacCatch
}
\description{
A data set containing year, catch and effort from longline to catch pelagic fish in eastern pacific sea (Schaefer, 1957).
}
\references{
Schaefer, Milner Baily. "A study of the dynamics of the fishery for yellow fin tuna in the eastern tropical Pacific Ocean." Bull. Inter-Amer. Trop. Tuna Comm. 2 (1957): 247-285.
}
\keyword{datasets}
