% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_goodcontrast.R
\docType{data}
\name{df.goodcontrast}
\alias{df.goodcontrast}
\title{Good contrast data with process and catch error}
\format{
A data frame with 20 rows and 3 variables:
\describe{
  \item{year}{year when fishing activity occured.}
  \item{catch}{catch per annum in MT.}
  \item{cpue}{catch per unit of effort per annum.}
  \item{biomass}{fish biomass per annum in MT.}
}
}
\usage{
df.goodcontrast
}
\description{
A data set containing year, catch, unit of effort and biomass which shows CPUE in good contrast.
This data shows a stochasticity using lognormal process and catch error.
}
\references{
Magnusson, Arni, and Ray Hilborn. "What makes fisheries data informative?." Fish and Fisheries 8, no. 4 (2007): 337-358.
}
\keyword{datasets}
