% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_goodcontrast0.R
\docType{data}
\name{df.goodcontrast0}
\alias{df.goodcontrast0}
\title{Good contrast data without process and catch error}
\format{
A data frame with 20 rows and 3 variables:
\describe{
  \item{year}{year when fishing activity occured.}
  \item{catch}{catch per annum in MT.}
  \item{effort}{effort unit per annum.}
}
}
\usage{
df.goodcontrast0
}
\description{
A data set containing year, catch and effort which shows CPUE in good contrast.
This data shows doesn't show stochasticity using lognormal process and catch error.
}
\references{
Magnusson, Arni, and Ray Hilborn. "What makes fisheries data informative?." Fish and Fisheries 8, no. 4 (2007): 337-358.
}
\keyword{datasets}
