% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_javaTrawl.R
\docType{data}
\name{javaTrawl}
\alias{javaTrawl}
\title{demersal trawl fishery data}
\format{
A data frame with 8 rows and 3 variables:
\describe{
  \item{year}{year when fishing activity occured.}
  \item{catch}{catch per annum in MT.}
  \item{effort}{effort unit per annum.}
}
}
\usage{
javaTrawl
}
\description{
A data set containing year, catch and effort from trawl to catch demersal fish in north java sea, Indonesia.
}
\references{
Dwiponggo, A. "Review of the demersal resources and fisheries in the Java Sea." Marine Fisheries Research Institute, Agency for Agricultural Research and Development, Ministry of Agriculture, Jakarta, Indonesia 36 (1979). Paper presented at th SCORRAD Meeting, 1979, Hong Kong.
}
\keyword{datasets}
