% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schaefer.R
\name{Par.init}
\alias{Par.init}
\title{Schaefer's data fitting and plotting}
\usage{
Par.init(K, B0, r, q, df, res = TRUE)
}
\arguments{
\item{K}{surplus production parameters which represents carrying capacity}

\item{B0}{surplus production parameters which represents biomass when no fishing
activity has started}

\item{r}{surplus production parameters which represents intrinsic growth rate}

\item{q}{surplus production parameters which represents catchability coefficient}

\item{df}{dataframe containing three columns; year, catch and effort}

\item{res}{option to show estimated data based on estimated surplus production input}
}
\description{
This function is used to predict the initial Biomass Dynamic Model parameters as input to do
the estimation in the latter process.

It also works to check the result of estimated parameters by plotting the data, whether the
data and estimated parameter is fitted.
}
\examples{

Par.init(K = 1000, B0 = 1000, r = 0.2, q = 0.00025, df = df.goodcontrast)

}
\references{
Hilborn, Ray, and Carl J. Walters, eds. Quantitative fisheries stock assessment: choice,
dynamics and uncertainty. Springer Science & Business Media, 1992.
}
