% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInit.R
\name{plotInit}
\alias{plotInit}
\title{Plotting catch, effort and CPUE}
\usage{
plotInit(df)
}
\arguments{
\item{df}{dataframe containing three columns; year, catch and effort}
}
\value{

}
\description{
This function plot the catch, effort and CPUE from a dataframe to check,
whether the plot shows one-way trip pattern or show a good contrast as required for the biomass dynamic model.
}
\examples{
plotInit(goodcontrast)

}
\references{
Hilborn, Ray, and Carl J. Walters, eds. Quantitative fisheries stock assessment: choice, dynamics and uncertainty. Springer Science & Business Media, 1992.

Magnusson, Arni, and Ray Hilborn. "What makes fisheries data informative?." Fish and Fisheries 8, no. 4 (2007): 337-358.
}
