% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schaefer.R
\name{run.Proj}
\alias{run.Proj}
\title{Schaefer's projection function}
\usage{
run.Proj(inpars, df, nyears, nsims, TAC = 1, plot = TRUE)
}
\arguments{
\item{inpars}{fitted surplus production parameters which consist of K (carrying capacity),
B0 (biomass when fishing is started), r (intrinsic growth rate), q (catchability coefficient), and
sigma (observation error)}

\item{df}{dataframe containing three columns; year, catch and effort}

\item{nyears}{number of years the projection for the fishery}

\item{nsims}{number of iteration performed}

\item{TAC}{number to drive the management level}

\item{plot}{whether the projection plot is produced or not}
}
\value{
The TAC is set at 1 in default. It would depend on the fishery to set whether the TAC can be set lower
to be more conservative or set higher to increase more catch.
For instance, when the TAC is set conservative at 0.8 of reference point, it will return
0.8MSY, and 0.8Emsy. In contrary, the TAC will increase catch when it is set at 1.2
(or other value higher than 1) and return 1.2MSY, and 1.2Emsy.
}
\description{
Function to create projection in Schaefer's model based on different reference points (MSY, Emsy) and
different Total Allowable Catch setting (TAC). This projection can be used to assist in the development of
limit and target reference point.
}
\examples{

fit <- calc.MSY(K=1000, B0=1000, r=0.2, q=0.00025, s.sigma=0.1,
df=df.goodcontrast, OWT=FALSE, currentF = 0.7, weight=0.5, plot=TRUE)

run.Proj(inpars=fit[[1]][,2], df=df.goodcontrast, nyears=30, nsims=100, TAC=1, plot=TRUE)


}
