% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SF_Eq.R
\name{SF_Eq}
\alias{SF_Eq}
\title{Schaefer and Fox' surplus production under equilibrium assumption}
\usage{
SF_Eq(df)
}
\arguments{
\item{df}{dataframe containing three columns; year, catch and effort}
}
\value{

}
\description{
This function calculate MSY and Emsy including its confident interval,
as well as the r squared and adj R squared, under equilibrium assumption using Schaefer and Fox models (Sparre and Venema, 1998).

This is just for educational purposes, do not use the result of this analysis for management purposes
as assumptions in this approach is rarely met and usually leads to potential yield and optimum effort overestimation (Hilborn and Walter, 1992; Polacheck, et al. 1993)
}
\examples{
SF_Eq(df.javaTrawl)

}
\references{
Sparre, P. and Venema, S.C. 1998. Introduction to tropical fish stock assessment, Part I: Manual. FAO Fisheries Technical Paper, 306, 1.

Hilborn, Ray, and Carl J. Walters, eds. Quantitative fisheries stock assessment: choice, dynamics and uncertainty. Springer Science & Business Media, 1992.

Polacheck, T., Hilborn, R., and A.E. Punt. 1993. Fitting surplus production models: Comparing methods and measuring uncertainty. Canadian Journal of Fisheries and Aquatic Sciences, 50: 2597-2607.
}
