% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFcalcRP.R
\name{SFcalcRP}
\alias{SFcalcRP}
\title{Schaefer and Fox' reference point calculator from data fitting parameters}
\usage{
SFcalcRP(inpars, SPmodel)
}
\arguments{
\item{inpars}{surplus production parameters which consist of K (carrying capacity), B0 (biomass when fishing is started), r (intrinsic growth rate), q (catchability coefficient)}

\item{SPmodel}{option on Surplus Production model; 1 for Schaefer and 2 for Fox}
}
\value{

}
\description{
This function is used to calculate the reference points (Bmsy, MSY, Emsy) under Schaefer and Fox model
}
\examples{
K <- 1000
B0 <- K
r <- 0.2
q <- 0.00025

inpars <- c(K, B0, r, q)
SFcalcRP(inpars=inpars, SPmodel=2)

}
\references{
Hilborn, Ray, and Carl J. Walters, eds. Quantitative fisheries stock assessment: choice, dynamics and uncertainty. Springer Science & Business Media, 1992.

Bonfil, R. (2005) ‘Fishery stock assessment models and their application to sharks’, in Musick, J. A. and Bonfil, Ramon (eds) Wildlife Conservation. Rome: Food and Agriculture Organizations of The United Nations, pp. 154–181.
}
