% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFpar_bayes.R
\name{SFpar_bayes}
\alias{SFpar_bayes}
\title{Schaefer and Fox' parameter estimation using bayesian}
\usage{
SFpar_bayes(df, K.prior = NULL, psi.prior = c(0.9, 0.25), r.prior, SPmodel)
}
\arguments{
\item{df}{dataframe containing three columns; year, catch and effort}

\item{K.prior}{carrying capacity (K) parameter prior. The default is NULL, can also receive input (mu, cv)}

\item{psi.prior}{ratio between B0/K parameter prior. The default is (0.9, 0.25), can also receive input (mu, cv)}

\item{r.prior}{intrinsic growth rate (r) parameter prior. The default is (mu, cv) with mu from rGrabber median parameter and cv is set at 0.5}

\item{SPmodel}{option on Surplus Production model; 1 for Schaefer and 2 for Fox}
}
\value{

}
\description{
Function to estimate the Schaefer and Fox' parameter using bayesian approach. This approach requires r from the assessed species as input prior in the bayesian,
to increase the accuracy of surplus production parameter estimation in Schaefer and Fox model.

This function requires JABBA package installed.
}
\examples{
# grab r intrinsic growth parameter and use it for the next analysis
r.val <- rGrabber("Hoplostethus atlanticus")

library("JABBA")
SFpar_bayes(df=df.eastpacCatch, K.prior, psi.prior, r.prior=r.val, SPmodel=1)

}
\references{
Winker, H., Carvalho, F., Kapur, M. (2018) JABBA: Just Another Bayesian Biomass Assessment. Fisheries Research 204: 275-288.
}
