% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smreg.R
\name{Smreg}
\alias{Smreg}
\title{Schaefer's multiple regression method}
\usage{
Smreg(df)
}
\arguments{
\item{df}{dataframe containing three columns; year, catch and effort}
}
\value{

}
\description{
This function calculate MSY and Emsy, as well as the r squared and adj R squared under non-equilibrium assumption.
The least square in the Walters and Hilborn (1976)'s multiple regression produces bias parameter (Uhler, 1979), which not advised to be used as it would result in further bias when calculating MSY and Emsy.
Based on the input, Hilborn and Walters (1992) adjusted the equation as input in the multiple regression.
}
\examples{
Smreg(df.eastpacCatch)

}
\references{
Carl J. Walters and Ray Hilborn. 1976. Adaptive Control of Fishing Systems. Journal of the Fisheries Research Board of Canada. 33(1): 145-159

Uhler, R. S. 1979. Least squares regression estimates of the Schaefer production model: some Monte Carlo simulation results. Can. J. Fish. Aquat. Sci. 37: 1284-1294.

Hilborn, Ray, and Carl J. Walters, eds. Quantitative fisheries stock assessment: choice, dynamics and uncertainty. Springer Science & Business Media, 1992.
}
