% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP_Eq.R
\name{SP_Eq}
\alias{SP_Eq}
\title{Surplus production under equilibrium assumption for Schaefer and Fox}
\usage{
SP_Eq(df)
}
\arguments{
\item{df}{dataframe containing three columns; year, catch and effort}
}
\value{

}
\description{
This function calculate MSY and fMSY including its confident interval,
as well as the r squared and adj R squared, under equilibrium assumption using Schaefer and Fox models
}
\examples{
SP_Eq(goodcontrast)

}
