% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schaefer.R
\name{Spar}
\alias{Spar}
\title{Schaefer's data fitting and plotting}
\usage{
Spar(inpars, df)
}
\arguments{
\item{inpars}{surplus production parameters which consist of K (carrying capacity), B0
(biomass when fishing is started), r (intrinsic growth rate), q (catchability coefficient)}

\item{df}{dataframe containing three columns; year, catch and effort}
}
\value{

}
\description{
This function is used to predict the initial Biomass Dynamic Model parameters as input to do
the estimation in the latter process.

It also works to check the result of estimated parameters by plotting the data, whether the
data and estimated parameter is fitted.
}
\examples{
K <- 1000
B0 <- K
r <- 0.2
q <- 0.00025

inpars <- c(K, B0, r, q)
Spar(inpars=inpars, df=df.goodcontrast)

}
\references{
Hilborn, Ray, and Carl J. Walters, eds. Quantitative fisheries stock assessment: choice,
dynamics and uncertainty. Springer Science & Business Media, 1992.
}
