% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spar_min.R
\name{Spar_min}
\alias{Spar_min}
\title{Schaefer's parameter estimation minimizing function}
\usage{
Spar_min(inpars, df, OWT = FALSE, Frate = 0.7, weight = 0.5)
}
\arguments{
\item{inpars}{surplus production parameters which consist of K (carrying capacity), B0
(biomass when fishing is started), r (intrinsic growth rate), q (catchability coefficient)}

\item{df}{dataframe containing three columns; year, catch and effort}

\item{OWT}{is CPUE plot showing One Way Trip pattern? The default is FALSE, but should be
replaced with TRUE when the plot shows One Way Trip}

\item{Frate}{exploitation rate collected from other survey. The default is 0.7}

\item{weight}{weight given to the deviation between observed and predicted value in
exploitation rate. The default is set at 0.5 with range between 0-1 (lower accuracy with high variance as closer to 0,
constrain the estimation procedure to fit the auxiliary information as closer to 1)}
}
\value{

}
\description{
Function used in the maximum likelihood minimization to estimate Schaefer's reference points.
Observation error is assumed to occur in the relationship between stock biomass and index of abundance,
therefore it is estimated assuming lognormal distribution in maximum likelihood (Polacheck et al., 1993).

Since fishing effort data collection are not always conducted regularly while catch is likely
have a better time series information, this function also allow for some lose of data.

This function also consider the different quality data, for instance if the data shows
a one way trip pattern which losing increasing rate of increase.
}
\examples{
K <- 1000
B0 <- K
r <- 0.2
q <- 0.00025

### Estimate parameters using optim
inpars <- c(log(K), log(B0), log(r), log(q), log(0.1))

fit <- optim(par=inpars,
             fn=Spar_min,
             df=df.goodcontrast,
             method="Nelder-Mead",
             OWT=FALSE, Frate = 0.7, weight = 0.5)

Spar_vals <- data.frame(SPpar = c("K", "B0", "r", "q", "sigma"),
                         init_pars = c(K, B0, r, q, 0.1),
                         fitted_pars = exp(fit$par))

}
\references{
Hilborn, Ray, and Carl J. Walters, eds. Quantitative fisheries stock assessment: choice,
dynamics and uncertainty. Springer Science & Business Media, 1992.

Polacheck, T., Hilborn, R., and A.E. Punt. 1993. Fitting surplus production models:
Comparing methods and measuring uncertainty. Canadian Journal of Fisheries and Aquatic Sciences, 50: 2597-2607.
}
