% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPprojS.R
\name{SPprojS}
\alias{SPprojS}
\title{Schaefer's projection function}
\usage{
SPprojS(inpars, df, nyears, TAC = 1, sigma = 1e-06)
}
\arguments{
\item{inpars}{fitted surplus production parameters which consist of K (carrying capacity), B0 (biomass when fishing is started), r (intrinsic growth rate), q (catchability coefficient)}

\item{df}{dataframe containing three columns; year, catch and effort}

\item{nyears}{number of years the projection for the fishery}

\item{sigma}{standard deviation estimated in the surplus production data fitting process}
}
\value{
sigma in default is set at a very low value (0.000001) to allow for a deterministic projection. In interest to create stochastic projection, the sigma value produced during maximum likelihood estimation or use another value can be utilized

The TAC (Total Allowable Catch) is set at 1 in default. It would depend on the fishery to set whether the TAC can be set lower to be more conservative or set higher to increase more catch.
}
\description{
Function to create projection in Schaefer's model based on different reference points.
}
\examples{
SPprojS(inpars=fitted_pars[1:4], df=goodcontrast, nyears=30, TAC=1, sigma=0.000001)

}
