% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_onewaytrip.R
\docType{data}
\name{df.onewaytrip}
\alias{df.onewaytrip}
\title{One way trip data with process and catch error}
\format{
A data frame with 20 rows and 3 variables:
\describe{
  \item{year}{year when fishing activity occured.}
  \item{catch}{catch per annum in MT.}
  \item{effort}{effort unit per annum.}
}
}
\usage{
df.onewaytrip
}
\description{
A data set containing year, catch and effort which shows CPUE in one way trip pattern.
This data shows stochasticity using lognormal process and catch error.
}
\references{
Magnusson, Arni, and Ray Hilborn. "What makes fisheries data informative?." Fish and Fisheries 8, no. 4 (2007): 337-358.
}
\keyword{datasets}
