% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGrabber.R
\name{rGrabber}
\alias{rGrabber}
\title{Grab intrinsic growth rate (r) parameter from Fishbase and SeaLifebase}
\usage{
rGrabber(SciName)
}
\arguments{
\item{SciName}{insert scientific name for the fish or aquatic biota}
}
\value{

}
\description{
Function to download intrinsic growth rate (r) parameter for specific species from Fishbase and SeaLifebase. The r growth parameter will serve as prior in the bayesian analysis in the surplus production model.

This function requires rfishpackage installed and internet connection.
}
\examples{
rgenerator("Sardinella lemuru")
rgenerator("Hoplostethus atlanticus")
rgenerator("Plectropomus areolatus")
rgenerator("Anabas testudineus")
rgenerator("Panulirus ornatus") #when stocks(SciName)==NA
rgenerator("Anguilla japonica")
rgenerator("Tegillarca granosa") #when estimate(SciName)==NA and stocks(SciName)==NA, but r information is available in the sealifebase web
}
\references{
https://www.fishbase.in/manual/key%20facts.htm
}
