% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGrabber.R
\name{rGrabber}
\alias{rGrabber}
\title{Grab intrinsic growth rate (r) parameter from Fishbase and SeaLifebase}
\usage{
rGrabber(SciName)
}
\arguments{
\item{SciName}{insert scientific name for the fish or aquatic biota}
}
\value{

}
\description{
Function to download intrinsic growth rate (r) parameter for specific species from Fishbase and SeaLifebase. The r growth parameter will serve as prior in the bayesian analysis in the surplus production model.

This function requires rfishpackage installed and internet connection.
}
\examples{
#grab r intrinsic growth rate parameter from the database
rGrabber("Sardinella lemuru") #for finfish species with a high r value, data grabbed from fishbase
rGrabber("Hoplostethus atlanticus") #for finfish species with a low r value, data grabbed from fishbase
rGrabber("Anguilla japonica") # for non finfish species, data grabbed from sealifebase

}
\references{
https://www.fishbase.in/manual/key%20facts.htm
}
