% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGrabber.R
\name{rGrabber}
\alias{rGrabber}
\title{Grab intrinsic growth rate (r) parameter from Fishbase and SeaLifebase}
\usage{
rGrabber(SciName)
}
\arguments{
\item{SciName}{insert scientific name for the fish or aquatic biota}
}
\value{

}
\description{
Function to download intrinsic growth rate (r) parameter for specific species from Fishbase and SeaLifebase. The r growth parameter will serve as prior in the bayesian analysis in the surplus production model.

This function requires rfishpackage installed and internet connection.
}
\examples{
#grab r intrinsic growth rate parameter from the database
rgenerator("Sardinella lemuru") #for finfish species with a high r value, data grabbed from fishbase
rgenerator("Hoplostethus atlanticus") #for finfish species with a high r value, data grabbed from fishbase
rgenerator("Anguilla japonica") # for non finfish species, data grabbed from sealifebase

# grab r intrinsic growth parameter and use it for the next analysis
r.val <- rgenerator("Hoplostethus atlanticus")

library(JABBA)
jbinput <- build_jabba(catch=df[,c(1,2)], cpue=data.frame(year=df[,1],cpue=df[,2]/df[,3]),
r.dist = c("lnorm", "range"), r.prior=c(r.val$median_r, 0.5), model.type = "Schaefer")

res.bayes <- fit_jabba(jbinput, init.r=r.val$median_r, quickmcmc=TRUE)

fitted_bayes <- c(res.bayes[["pars"]]$Median[1], # K
                  res.bayes[["pars"]]$Median[4]*res.bayes[["pars"]]$Median[1], #B0
                  res.bayes[["pars"]]$Median[2], # r
                  res.bayes[["pars"]]$Median[3]) #q

}
\references{
https://www.fishbase.in/manual/key%20facts.htm
}
